requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.7.0

# SharePoint integration
Office365-REST-Python-Client>=2.5.0

# Logging and monitoring
structlog>=23.1.0
colorama>=0.4.6

# Date and time handling
python-dateutil>=2.8.2

# Configuration management
pyyaml>=6.0

# HTTP client with better error handling
httpx>=0.24.0

# Development dependencies (uncomment when needed)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.0.0
# mypy>=1.5.0
